+++ app/src/main/java/org/breezyweather/sources/CommonConverter.kt
@@ -579,7 +579,11 @@ fun completeDailyListFromHourlyList(
             uV = if (daily.uV?.index != null) daily.uV else getDailyUVFromHourlyList(
                 hourlyListByDay.getOrElse(theDayFormatted) { null }
             ),
-            hoursOfSun = daily.hoursOfSun ?: getHoursOfDay(newSun.riseDate, newSun.setDate)
+            hoursOfSun = daily.hoursOfSun ?: getHoursOfDay(
+                hourlyListByDay.getOrElse(theDayFormatted) { null },
+                newSun.riseDate,
+                newSun.setDate
+            )
         )
     }
 }
@@ -1131,7 +1135,26 @@ private fun getDailyUVFromHourlyList(hourlyList: List<HourlyWrapper>? = null): U
     return UV(index = hourlyListWithUV.max())
 }
 
-private fun getHoursOfDay(sunrise: Date?, sunset: Date?): Double? {
+/**
+ * Returns the sunshine duration if present in hourlyList,
+ * otherwise fallback to hours of sun using sunrise/set
+ *
+ * @param hourlyList hourly list for that day
+ * @param sunrise sun rise date for that day
+ * @param sunset sun set date for that day
+ */
+private fun getHoursOfDay(
+    hourlyList: List<HourlyWrapper>?,
+    sunrise: Date?,
+    sunset: Date?
+): Double? {
+    if (hourlyList != null) {
+        val hourlyWithSunshine = hourlyList.filter { it.sunshineDuration != null }
+        if (hourlyWithSunshine.isNotEmpty()) {
+            return hourlyWithSunshine.sumOf { it.sunshineDuration!! }
+        }
+    }
+
     return if (sunrise == null || sunset == null) {
         // Polar night
         0.0