+++ app/src/main/java/org/breezyweather/sources/openmeteo/OpenMeteoService.kt
@@ -102,6 +102,7 @@ class OpenMeteoService @Inject constructor(
             "apparent_temperature_min",
             "sunrise",
             "sunset",
+            "sunshine_duration",
             "uv_index_max"
         )
         val hourly = arrayOf(